Moves faker function and method to lang package
