+++ internal/bloblang/query/methods_structured.go
@@ -9,7 +9,6 @@ import (
 	"strings"
 
 	"github.com/Jeffail/gabs/v2"
-	"github.com/bxcodec/faker/v3"
 	jsonschema "github.com/xeipuuv/gojsonschema"
 )
 
@@ -1562,65 +1561,4 @@ func mapWithout(m map[string]interface{}, paths [][]string) map[string]interface
 		}
 	}
 	return newMap
-}
-
-//------------------------------------------------------------------------------
-
-var _ = registerSimpleMethod(
-	NewMethodSpec(
-		"fake",
-		"Uses the values in the structure to map to functions in the faker library. If value doesn't match a faker method, it remains unchanged.",
-	).InCategory(
-		MethodCategoryObjectAndArray, "",
-		NewExampleSpec(``,
-			`root.result = this.fake()`,
-		),
-	),
-	func(*ParsedParams) (simpleMethod, error) {
-		return func(v interface{}, ctx FunctionContext) (interface{}, error) {
-			return populateFakeData(v), nil
-		}, nil
-	},
-)
-
-func populateFakeData(data interface{}) interface{} {
-	str, isString := data.(string)
-	if isString {
-		fakeValue := GetFakeValue(str)
-		return fakeValue
-	}
-
-	array, isArray := data.([]interface{})
-	if isArray {
-		var newArray []interface{}
-		for _, value := range array {
-			newValue := populateFakeData(value)
-			newArray = append(newArray, newValue)
-		}
-		return newArray
-	}
-
-	m, isMap := data.(map[string]interface{})
-	if isMap {
-		newMap := map[string]interface{}{}
-		for key, value := range m {
-			newValue := populateFakeData(value)
-			newMap[key] = newValue
-		}
-		return newMap
-	}
-
-	return data
-}
-
-// GetFakeValue returns fake data generated by the faker function corresponding to the input string
-func GetFakeValue(value string) string {
-	result := value
-	switch strings.ToLower(value) {
-	case "email":
-		result = faker.Email()
-	case "date":
-		result = faker.Date()
-	}
-	return result
 }
