+++ website/docs/guides/bloblang/functions.md
@@ -9,7 +9,7 @@ description: A list of Bloblang functions
 
      To make changes please edit the contents of:
      internal/bloblang/query/functions.go
-     internal/docs/bloblang.go
+     internal/docs/bloblang_methods.go
 -->
 
 import Tabs from '@theme/Tabs';
@@ -76,22 +76,6 @@ root.new_nums = this.nums.map_each(num -> if num < 10 { deleted() } else { num -
 
 # In:  {"nums":[3,11,4,17]}
 # Out: {"new_nums":[1,7]}
-```
-
-### `fake`
-
-The `fake` takes in a string and returns a generated string using the faker function specified
-
-#### Parameters
-
-**`function`** &lt;string&gt; An identifier for the faker function to use.  
-
-#### Examples
-
-
-```coffee
-root = this
-root.email = fake("email")
 ```
 
 ### `ksuid`
